+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.6 / 2020-02-27
+- [FEATURE] Finished fadeTo effect.
+- [TESTS] Added testing for effects.
+
 ## 0.0.5 / 2020-02-26
 - [FEATURE] Finished panTo effect.
 - [HOTFIX] Fixed issue with zooming that caused it to finish before the duration of the effect had been reached.